+++ app/components/primer/primer.ts
@@ -1,2 +1,3 @@
+import './auto_complete_component'
 import './tab_container_component'
 import './time_ago_component'